test(outfitter): split heavy init and upgrade integration suites
